requirements.txt
anyio==3.6.1
APScheduler==3.9.1
asgiref==3.5.2
attrs==21.4.0
beautifulsoup4==4.11.1
beautifultable==0.7.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-bootstrap4==22.1
django-filter==22.1
django-tables2==2.4.1
djangorestframework==3.13.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
install==1.3.5
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
model-bakery==1.6.0
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-env==0.6.2
python-telegram-bot==20.0a2
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
tornado==6.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
whitenoise==6.2.0
zipp==3.8.1
